agent/requirements.txt
# OTTO LiveKit Agent Dependencies

livekit-agents[silero,turn-detector]>=1.0.0
livekit-plugins-anthropic>=0.2.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-cartesia>=0.4.0
livekit-plugins-noise-cancellation>=0.2.0

python-dotenv>=1.0.0
httpx>=0.27.0
